fixed abort when 2 processes run at once. check whether filtered file exists just before renaming
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder/trunk/moses@281 1f5c12ca-751b-0410-a591-d2e778427230

+++ src/PhraseDictionary.cpp
@@ -150,14 +150,17 @@ void PhraseDictionary::Load(const std::vector<FactorType> &input
 	{
 		tempFile.close();
 		using namespace boost::filesystem;
-		try 
+		if (!exists(path(hashFilePath, native)))
 		{
-			rename( path(tempFilePath, native) , path(hashFilePath, native) );
-		}
-		catch (...)
-		{ // copy instead
-			copy_file(path(tempFilePath, native) , path(hashFilePath, native) );
-			remove(tempFilePath);
+			try 
+			{
+				rename( path(tempFilePath, native) , path(hashFilePath, native) );
+			}
+			catch (...)
+			{ // copy instead
+				copy_file(path(tempFilePath, native) , path(hashFilePath, native) );
+				remove(tempFilePath);
+			}
 		}
 #ifndef _WIN32
 		// change permission to let everyone use cached file
